# ui/openapi.json
{
  "components": {
    "schemas": {
      "AcceptIn": {
        "properties": {
          "edits": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edits"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "AcceptIn",
        "type": "object"
      },
...
      "GraphEdgeOut": {
        "description": "One time segment of one predicate — never a collapsed span (ADR-030).\n\nThere is no ``weight``. ``support`` carries each visible claim's three\ngrading dimensions and the corroboration/contradiction counts around them,\nso a reader can reach the evidence instead of trusting a scalar.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "record_count": {
            "title": "Record Count",
            "type": "integer"
          },
          "segment_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment From"
          },
          "segment_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment To"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "support": {
            "additionalProperties": true,
            "title": "Support",
            "type": "object"
          }
        },
        "required": [
          "edge_id",
          "subject_id",
          "object_id",
          "predicate",
          "category",
          "segment_from",
          "segment_to",
          "record_count",
          "support"
        ],
        "title": "GraphEdgeOut",
        "type": "object"
      },
      "GraphExpandIn": {
        "description": "A bounded traversal request (specs/06 §2.6).\n\nEvery bound is clamped rather than rejected (specs/06 §4): a client asking\nfor six hops gets three and is told the result was truncated, which is more\nuseful than a 422 that teaches nothing about the limit.",
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "max_elements": {
            "default": 2000,
            "minimum": 1.0,
            "title": "Max Elements",
            "type": "integer"
          },
          "max_hops": {
            "default": 1,
            "minimum": 0.0,
            "title": "Max Hops",
            "type": "integer"
          },
          "seed_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Seed Ids",
            "type": "array"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "title": "GraphExpandIn",
        "type": "object"
      },
      "GraphNodeOut": {
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "label",
          "entity_type"
        ],
        "title": "GraphNodeOut",
        "type": "object"
      },
      "GraphPathOut": {
        "properties": {
          "edge_ids": {
            "items": {
              "type": "string"
            },
            "title": "Edge Ids",
            "type": "array"
          },
          "entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          }
        },
        "required": [
          "entity_ids",
          "edge_ids"
        ],
        "title": "GraphPathOut",
        "type": "object"
      },
      "GraphPathsIn": {
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "from_id": {
            "title": "From Id",
            "type": "string"
          },
          "max_hops": {
            "default": 5,
            "minimum": 1.0,
            "title": "Max Hops",
            "type": "integer"
          },
          "max_paths": {
            "default": 25,
            "minimum": 1.0,
            "title": "Max Paths",
            "type": "integer"
          },
          "to_id": {
            "title": "To Id",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "from_id",
          "to_id"
        ],
        "title": "GraphPathsIn",
        "type": "object"
      },
      "GraphPathsOut": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeOut"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/GraphPathOut"
            },
            "title": "Paths",
            "type": "array"
          },
          "resolved_seed_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resolved Seed Ids",
            "type": "array"
          },
          "seed_ids": {
            "items": {
              "type": "string"
            },
            "title": "Seed Ids",
            "type": "array"
          },
          "stamps": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectionStampsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "nodes",
          "edges",
          "seed_ids",
          "resolved_seed_ids",
          "truncated",
          "stamps",
          "paths"
        ],
        "title": "GraphPathsOut",
        "type": "object"
      },
      "GraphViewOut": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeOut"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          },
          "resolved_seed_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resolved Seed Ids",
            "type": "array"
          },
          "seed_ids": {
            "items": {
              "type": "string"
            },
            "title": "Seed Ids",
            "type": "array"
          },
          "stamps": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectionStampsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "nodes",
          "edges",
          "seed_ids",
          "resolved_seed_ids",
          "truncated",
          "stamps"
        ],
        "title": "GraphViewOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectionStampsOut": {
        "description": "Which build produced these rows, and whether it is behind (specs/06 §3).",
        "properties": {
          "active_revision_id": {
            "title": "Active Revision Id",
            "type": "integer"
          },
          "builder_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Builder Version"
          },
          "built_at_revision_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built At Revision Id"
          },
          "ontology_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ontology Version"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "built_at_revision_id",
          "active_revision_id",
          "ontology_version",
          "builder_version",
          "stale"
        ],
        "title": "ProjectionStampsOut",
        "type": "object"
      },
      "RejectIn": {
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectIn",
        "type": "object"
      },
...
      "SuggestionOut": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "producer_meta": {
            "additionalProperties": true,
            "title": "Producer Meta",
            "type": "object"
          },
          "producer_version": {
            "title": "Producer Version",
            "type": "string"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "result_claim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Claim Id"
          },
          "result_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Decision Id"
          },
          "result_relation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Relation"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suggestion_id": {
            "title": "Suggestion Id",
            "type": "string"
          },
          "suggestion_kind": {
            "title": "Suggestion Kind",
            "type": "string"
          },
          "target_action": {
            "title": "Target Action",
            "type": "string"
          }
        },
        "required": [
          "suggestion_id",
          "suggestion_kind",
          "schema_version",
          "payload",
          "target_action",
          "producer",
          "producer_version",
          "producer_meta",
          "record_id",
          "case_id",
          "status",
          "decided_by",
          "decided_at",
          "decision_note",
          "result_claim_id",
          "result_decision_id",
          "result_relation",
          "created_at"
        ],
        "title": "SuggestionOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Governed claims-based intelligence platform (speckit Phase 2).",
    "title": "Aegis API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/graph/expand": {
      "post": {
        "description": "Breadth-first expansion from seeds, or the bounded overview without them.\n\nPOST rather than GET because the request is a query object — seed sets,\ncategory filters and time windows — not a resource address, and putting a\nhundred entity ids in a query string is how URL-length limits become silent\ntruncation of an authorization-relevant input.",
        "operationId": "expandGraph",
        "parameters": [
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphExpandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphViewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Expand Graph",
        "tags": [
          "graph"
        ]
      }
    },
    "/v1/graph/paths": {
      "post": {
        "description": "Shortest routes between two entities, bounded by hops and count.\n\nA path here is a claim about *what the evidence connects*, not about what\nanyone did: every edge on it opens onto its claims, and none of them is\ngraded by its position on the path (Article IX).",
        "operationId": "graphPaths",
        "parameters": [
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphPathsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphPathsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Graph Paths",
        "tags": [
          "graph"
        ]
      }
    },
    "/v1/review-queue": {
      "get": {
        "operationId": "listSuggestions",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "producer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Producer"
            }
          },
          {
            "in": "query",
            "name": "record",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SuggestionOut"
                  },
                  "title": "Response Listsuggestions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Suggestions",
        "tags": [
          "review-queue"
        ]
      }
    },
    "/v1/review-queue/{suggestion_id}/accept": {
      "post": {
        "operationId": "acceptSuggestion",
        "parameters": [
          {
            "in": "path",
            "name": "suggestion_id",
            "required": true,
            "schema": {
              "title": "Suggestion Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Suggestion",
        "tags": [
          "review-queue"
        ]
      }
    },
    "/v1/review-queue/{suggestion_id}/reject": {
      "post": {
        "operationId": "rejectSuggestion",
        "parameters": [
          {
            "in": "path",
            "name": "suggestion_id",
            "required": true,
            "schema": {
              "title": "Suggestion Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Suggestion",
        "tags": [
          "review-queue"
        ]
      }
    },
...
  }
}